fix: shiki rendering, refactor: dev server, feat: revalidate on demand
