+++ packages/zzap/package.json
@@ -11,6 +11,7 @@
   },
   "scripts": {
     "dev": "bun package.tsx dev",
+    "debug": "bun run dev",
     "package": "bun package.tsx package",
     "bump": "bun ./scripts/bump.ts",
     "distribute": "npm publish",
