+++ sites/zzap.dev/docs/src/types/DynamicData.ts
