+++ sites/zzap.dev/package.json
@@ -3,7 +3,8 @@
   "description": "Documentation website for zzap",
   "authors": "René-Pier Deshaies-Géllinas",
   "scripts": {
-    "dev": "zzap watch --port 3000 --debug",
+    "dev": "zzap watch --port 3000",
+    "debug": "bun run dev --debug",
     "start": "zzap build --debug && zzap start --port 3000 --debug",
     "build": "zzap build --debug"
   },
